done some changes

+++ src/context/PcContextProvider.jsx
@@ -1,17 +1,21 @@
-import { addDoc, collection, deleteDoc, doc, getDocs, updateDoc } from 'firebase/firestore';
-import React, { createContext, useEffect, useState } from 'react'
+import { addDoc, collection, deleteDoc, doc, getDocs, increment, updateDoc } from 'firebase/firestore';
+import React, { createContext, useContext, useEffect, useState } from 'react'
 import { toast } from 'react-toastify';
 import { db } from '../config/firebase';
+import { LabContext } from './LabContextProvider';
+import { StudentContext } from './StudentContextProvider';
 
 export const PcContext = createContext();
 
 const PcContextProvider = ({ children }) => {
     const [pcs, setPcs] = useState([]);
     const [editId, setEditId] = useState(null);
-    
+    const { fetchData } = useContext(LabContext);
+    const { students } = useContext(StudentContext);
+
     useEffect(() => {
         fetchPc();
-    }, [])
+    }, [students])
 
 
     const addPc = async (pc) => {
@@ -20,6 +24,8 @@ const PcContextProvider = ({ children }) => {
                 createdAt: new Date(),
                 ...pc
             })
+            await updateDoc(doc(db, "labs", pc.labId), { initialCapacity: increment(-1) });
+            fetchData();
             fetchPc();
         } catch (error) {
             toast.error("Something Went Wrong")
@@ -33,7 +39,7 @@ const PcContextProvider = ({ children }) => {
             const pcList = docs.map((pc) => {
                 return {
                     pcId: pc.id,
-                    ...pc.data()
+                    ...pc.data(),
                 }
             });
             setPcs(pcList);
@@ -43,9 +49,26 @@ const PcContextProvider = ({ children }) => {
     }
 
     const deletePc = async (id) => {
+        const pc = pcs?.find((pc) => pc.pcId === id)
+        const stu = students?.find((std) => {
+            return std.pcId === id;
+        });
+
+        if (!pc) {
+            toast.error("PC not found");
+            return;
+        }
+
         try {
             await deleteDoc(doc(db, "pcs", id));
+            await updateDoc(doc(db, "labs", pc.labId), { initialCapacity: increment(1) });
+
+            if (stu) {
+                await updateDoc(doc(db, "students", stu.id), { pcId: null });
+            }
+
             fetchPc();
+            fetchData();
             toast.success("Pc Deleted Successfully...");
         } catch (error) {
             toast.error("Something Went Wrong...");
@@ -53,16 +76,19 @@ const PcContextProvider = ({ children }) => {
     }
 
     const updatePc = async (input) => {
-            if(!editId) return;
-            try {
-                await updateDoc(doc(db, "pcs", editId), input)
-                toast.success("PC updated successfully...");
-                fetchPc();
-                setEditId(null);
-            } catch (error) {
-                toast.error("Something went wrong...");
-            }
-        
+        if (!editId) {
+            toast.error("No PC selected for update")
+            return
+        };
+        try {
+            await updateDoc(doc(db, "pcs", editId), input)
+            toast.success("PC updated successfully...");
+            fetchPc();
+            setEditId(null);
+        } catch (error) {
+            toast.error("Something went wrong...");
+        }
+
     }
 
     const data = { addPc, pcs, deletePc, editId, setEditId, updatePc }
